move sovereign cloud logic to own component
