+++ src/messages/GE.json
@@ -352,5 +352,6 @@
   "and experiment on": " and experiment on",
   "Scope consent failed": "Scope consent failed",
   "Select cloud": "Select cloud",
-  "Cloud selected": "Cloud selected"
+  "Cloud selected": "Cloud selected",
+  "You have access to sovereign clouds": "You have access to sovereign clouds"
 }
